% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-fancy.R
\name{print_fancy}
\alias{print_fancy}
\title{Print Fancy Text}
\usage{
print_fancy(x)
}
\arguments{
\item{x}{A length one character vector}
}
\value{
x, invisibly
}
\description{
This function prints the string provided to it and formats it to be bold with
a green check mark
}
\examples{
print_fancy("test")
}
